bug-hunter/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
Pillow==10.1.0
pytesseract==0.3.10
pylint==3.0.3
sentence-transformers==2.2.2
pinecone-client==3.0.0
aiofiles==23.2.1
